+++ CHANGELOG.md
@@ -74,6 +74,10 @@ Current
     * Created `LegacyGenerator` as a bridge interface from the existing constructor based api request impls and the factory based value object usage.
 
 ### Added:
+- [Add logging for cache miss ,hit and potential hits](https://github.com/yahoo/fili/pull/1115)
+   * Utilize `BardCacheInfo` to add logging for cache hit and miss.
+   * Re-sequence in `CacheV2ResponseProcessor' to send response back to the user first and then set cache.
+   
 - [Add canonicalization of ArrayNode in cache key](https://github.com/yahoo/fili/pull/1112)
    * Added parsing and canonicalization of ArrayNodes in cache key generation.